+++ packages/web-framework/types/bin/web-server.d.ts
@@ -1,3 +1,4 @@
+/// <reference types="node" />
 export = TiWebServer;
 import ServiceConsumer = require("@ti-engine/core/service-consumer");
 export type NodeServer = import("node:http").Server;